+++ .changeset/calm-bears-wave.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': major
+---
+
+breaking: move `RequestEvent` and `Cookies` to `$app/server`
