+++ documentation/docs/20-core-concepts/20-load.md
@@ -292,7 +292,7 @@ export async function load({ fetch, params }) {
 
 ## Cookies
 
-A server `load` function can get [`cookies`](@sveltejs-kit#Cookies) as shown below. When setting cookies, SvelteKit provides default values for `httpOnly`, `secure`, and `path` — as described in [the API documentation](@sveltejs-kit#Cookies) — in order to improve security and developer experience.
+A server `load` function can get [`cookies`]($app-server#Cookies) as shown below. When setting cookies, SvelteKit provides default values for `httpOnly`, `secure`, and `path` — as described in [the API documentation]($app-server#Cookies) — in order to improve security and developer experience.
 
 ```js
 /// file: src/routes/+layout.server.js
