+++ documentation/docs/60-appendix/35-migrating-to-sveltekit-3.md
@@ -286,6 +286,10 @@ Types like `EnvVarConfig`, used with [`defineEnvVars`](@sveltejs-kit-env#defineE
 
 Types like `Handle`, which defines the type of your [`handle`](hooks#handle) hook, now live in `@sveltejs/kit/hooks`.
 
+### Request-related types moved to `$app/server`
+
+Types like `RequestEvent` and `Cookies`, used by [`getRequestEvent`]($app-server#getRequestEvent), now live in `$app/server`.
+
 ## `@sveltejs/kit/hooks`
 
 The `defineEnvVars` function has moved from `@sveltejs/kit/hooks` to `@sveltejs/kit/env`.
@@ -355,7 +359,7 @@ SvelteKit now uses [`cookie`](https://npmx.dev/package/cookie) v2, which involve
 When setting a cookie without an explicit `path` (which was previously forbidden), the path defaults to `'/'` rather than the current request path, meaning the cookie applies to the entire site. This matches what most developers expect. You can pass an explicit `path` if necessary:
 
 ```ts
-declare const cookies: import('@sveltejs/kit').Cookies;
+declare const cookies: import('$app/server').Cookies;
 declare const name: string;
 declare const value: string;
 // ---cut---
