+++ packages/kit/src/core/sync/write_types/index.js
@@ -354,7 +354,9 @@ function update_types(config, routes, route, root, to_delete = new Set()) {
 	}
 
 	if (route.leaf?.server || route.layout?.server || route.endpoint) {
-		exports.push('export type RequestEvent = Kit.RequestEvent<RouteParams, RouteId>;');
+		exports.push(
+			"export type RequestEvent = import('$app/server').RequestEvent<RouteParams, RouteId>;"
+		);
 	}
 
 	const output = [imports.join('\n'), declarations.join('\n'), exports.join('\n')]
