+++ packages/kit/src/exports/hooks/public.d.ts
@@ -1,5 +1,6 @@
 import { StandardSchemaV1 } from '@standard-schema/spec';
-import { NavigationEvent, RequestEvent } from '@sveltejs/kit';
+import { NavigationEvent } from '@sveltejs/kit';
+import { RequestEvent } from '$app/server';
 import { MaybePromise } from 'types';
 
 export * from './index.js';
