+++ packages/kit/src/exports/vite/dev/index.js
@@ -1,4 +1,5 @@
-/** @import { RequestEvent, SSRManifest } from '@sveltejs/kit' */
+/** @import { SSRManifest } from '@sveltejs/kit' */
+/** @import { RequestEvent } from '$app/server' */
 /** @import { EnvironmentModuleNode, ErrorPayload, ResolvedConfig, ViteDevServer } from 'vite' */
 /** @import { ManifestData, PrerenderOption, RemoteChunk, ServerModule, SSRNode, UniversalNode, ValidatedConfig } from 'types' */
 import process from 'node:process';
