+++ packages/kit/src/runtime/server/cookie.js
@@ -76,11 +76,11 @@ export function get_cookies(request, url) {
 		secure: !__SVELTEKIT_DEV__ && !(url.hostname === 'localhost' && url.protocol === 'http:')
 	};
 
-	/** @type {import('@sveltejs/kit').Cookies} */
+	/** @type {import('$app/server').Cookies} */
 	const cookies = {
 		// The JSDoc param annotations appearing below for get, set and delete
 		// are necessary to expose the `cookie` library types to
-		// typescript users. `@type {import('@sveltejs/kit').Cookies}` above is not
+		// typescript users. `@type {import('$app/server').Cookies}` above is not
 		// sufficient to do so.
 
 		get(name, opts) {
