+++ packages/kit/src/runtime/server/fetch.js
@@ -8,7 +8,7 @@ import { fork_state_for_subrequest } from './state.js';
 
 /**
  * @param {{
- *   event: import('@sveltejs/kit').RequestEvent;
+ *   event: import('$app/server').RequestEvent;
  *   options: import('types').SSROptions;
  *   manifest: import('@sveltejs/kit').SSRManifest;
  *   state: import('types').RequestState;
