+++ packages/kit/src/runtime/server/page/actions.js
@@ -1,4 +1,5 @@
-/** @import { RequestEvent, Actions } from '@sveltejs/kit' */
+/** @import { Actions } from '@sveltejs/kit' */
+/** @import { RequestEvent } from '$app/server' */
 /** @import { ActionResult } from '$app/forms' */
 /** @import { SSROptions, SSRNode, ServerNode } from 'types' */
 import { DEV } from 'esm-env';
