+++ packages/kit/src/runtime/server/page/load_data.spec.js
@@ -2,7 +2,7 @@ import { assert, expect, test } from 'vitest';
 import { create_universal_fetch } from './load_data.js';
 
 /**
- * @param {Partial<Pick<import('@sveltejs/kit').RequestEvent, 'fetch' | 'url' | 'request' | 'route'>>} event
+ * @param {Partial<Pick<import('$app/server').RequestEvent, 'fetch' | 'url' | 'request' | 'route'>>} event
  */
 function create_fetch(event) {
 	// @ts-expect-error
@@ -15,7 +15,7 @@ function create_fetch(event) {
 	// @ts-expect-error
 	event.url ||= new URL('https://domain-a.com');
 	return create_universal_fetch(
-		/** @type {Pick<import('@sveltejs/kit').RequestEvent, 'fetch' | 'url' | 'request' | 'route'>} */ (
+		/** @type {Pick<import('$app/server').RequestEvent, 'fetch' | 'url' | 'request' | 'route'>} */ (
 			event
 		),
 		undefined,
