+++ packages/kit/src/types/internal.d.ts
@@ -8,11 +8,10 @@ import {
 	Server,
 	ServerInitOptions,
 	Actions,
-	RequestEvent,
 	SSRManifest,
 	Emulator
 } from '@sveltejs/kit';
-import { RemoteFormIssue, RemoteQuery, RemoteLiveQuery } from '$app/server';
+import { RemoteFormIssue, RemoteQuery, RemoteLiveQuery, RequestEvent } from '$app/server';
 import {
 	ClientInit,
 	Handle,
